+++ src/main/java/ru/grig/ratingRestaurant/util/ValidationUtil.java
@@ -5,12 +5,12 @@ import ru.grig.ratingRestaurant.util.exception.NotFoundException;
 
 public class ValidationUtil {
 
-    public static <T> T checkNotFoundWithId(T object, long id) {
+    public static <T> T checkNotFoundWithId(T object, int id) {
         checkNotFound(object != null, "id = " + id);
         return object;
     }
 
-    public static void checkNotFoundWithId(boolean found, long id) {
+    public static void checkNotFoundWithId(boolean found, int id) {
         checkNotFound(found, "id=" + id);
     }
 
@@ -31,7 +31,7 @@ public class ValidationUtil {
         }
     }
 
-    public static void assureIdConsistent(AbstractBaseEntity entity, long id) {
+    public static void assureIdConsistent(AbstractBaseEntity entity, int id) {
 //      conservative when you reply, but accept liberally (http://stackoverflow.com/a/32728226/548473)
         if (entity.isNew()) {
             entity.setId(id);
